feat(starfish): Fixes spans view operation, action, and domain filter dropdown items to match actual results (#52659)
Dropdown options for operation, action, and domain filter were not
matching the actual unique values in the table result. This is because
the query filters for the operation, action, and domain requests didnt
match the filter for the table request.
