+++ static/app/views/starfish/queries/useSpanList.tsx
@@ -7,26 +7,11 @@ import type {Sort} from 'sentry/utils/discover/fields';
 import {DiscoverDatasets} from 'sentry/utils/discover/types';
 import {useLocation} from 'sentry/utils/useLocation';
 import {ModuleName, SpanMetricsFields} from 'sentry/views/starfish/types';
+import {buildEventViewQuery} from 'sentry/views/starfish/utils/buildEventViewQuery';
 import {useWrappedDiscoverQuery} from 'sentry/views/starfish/utils/useSpansQuery';
-import {NULL_SPAN_CATEGORY} from 'sentry/views/starfish/views/webServiceView/spanGroupBreakdownContainer';
 
-const {
-  SPAN_SELF_TIME,
-  SPAN_DESCRIPTION,
-  SPAN_GROUP,
-  SPAN_OP,
-  SPAN_DOMAIN,
-  SPAN_ACTION,
-  SPAN_MODULE,
-} = SpanMetricsFields;
-
-const SPAN_FILTER_KEYS = [
-  SPAN_OP,
-  SPAN_DOMAIN,
-  SPAN_ACTION,
-  `!${SPAN_MODULE}`,
-  '!span.category',
-];
+const {SPAN_SELF_TIME, SPAN_DESCRIPTION, SPAN_GROUP, SPAN_OP, SPAN_DOMAIN} =
+  SpanMetricsFields;
 
 export type SpanMetrics = {
   'http_error_count()': number;
@@ -84,13 +69,13 @@ function getEventView(
   spanCategory?: string,
   sorts?: Sort[]
 ) {
-  const query = buildEventViewQuery(
+  const query = buildEventViewQuery({
     moduleName,
     location,
     transaction,
     method,
-    spanCategory
-  )
+    spanCategory,
+  })
     .filter(Boolean)
     .join(' ');
 
@@ -130,60 +115,4 @@ function getEventView(
   }
 
   return eventView;
-}
-
-function buildEventViewQuery(
-  moduleName: ModuleName,
-  location: Location,
-  transaction?: string,
-  method?: string,
-  spanCategory?: string
-) {
-  const {query} = location;
-  const result = Object.keys(query)
-    .filter(key => SPAN_FILTER_KEYS.includes(key))
-    .filter(key => Boolean(query[key]))
-    .map(key => {
-      const value = query[key];
-      const isArray = Array.isArray(value);
-
-      if (key === '!span.category' && isArray && value.includes('db')) {
-        // When omitting database spans, explicitly allow `db.redis` spans, because
-        // we're not including those spans in the database category
-        const categoriesAsideFromDatabase = value.filter(v => v !== 'db');
-        return `(!span.category:db OR ${SPAN_OP}:db.redis) !span.category:[${categoriesAsideFromDatabase.join(
-          ','
-        )}]`;
-      }
-
-      return `${key}:${isArray ? `[${value}]` : value}`;
-    });
-
-  result.push(`has:${SPAN_DESCRIPTION}`);
-
-  if (moduleName !== ModuleName.ALL) {
-    result.push(`${SPAN_MODULE}:${moduleName}`);
-  }
-
-  if (moduleName === ModuleName.DB) {
-    result.push(`!${SPAN_OP}:db.redis`);
-  }
-
-  if (defined(spanCategory)) {
-    if (spanCategory === NULL_SPAN_CATEGORY) {
-      result.push(`!has:span.category`);
-    } else if (spanCategory !== 'Other') {
-      result.push(`span.category:${spanCategory}`);
-    }
-  }
-
-  if (transaction) {
-    result.push(`transaction:${transaction}`);
-  }
-
-  if (method) {
-    result.push(`transaction.method:${method}`);
-  }
-
-  return result;
 }
