+++ static/app/views/starfish/views/spans/selectors/domainSelector.tsx
@@ -1,6 +1,7 @@
 import {ReactNode} from 'react';
 import {browserHistory} from 'react-router';
 import {Location} from 'history';
+import omit from 'lodash/omit';
 
 import {CompactSelect} from 'sentry/components/compactSelect';
 import {t} from 'sentry/locale';
@@ -8,10 +9,10 @@ import EventView from 'sentry/utils/discover/eventView';
 import {DiscoverDatasets} from 'sentry/utils/discover/types';
 import {useLocation} from 'sentry/utils/useLocation';
 import {ModuleName, SpanMetricsFields} from 'sentry/views/starfish/types';
+import {buildEventViewQuery} from 'sentry/views/starfish/utils/buildEventViewQuery';
 import {useSpansQuery} from 'sentry/views/starfish/utils/useSpansQuery';
-import {NULL_SPAN_CATEGORY} from 'sentry/views/starfish/views/webServiceView/spanGroupBreakdownContainer';
 
-const {SPAN_MODULE, SPAN_OP, SPAN_DESCRIPTION, SPAN_DOMAIN} = SpanMetricsFields;
+const {SPAN_DOMAIN} = SpanMetricsFields;
 
 type Props = {
   moduleName?: ModuleName;
@@ -69,29 +70,18 @@ const LABEL_FOR_MODULE_NAME: {[key in ModuleName]: ReactNode} = {
   '': t('Domain'),
 };
 
-function getEventView(location: Location, moduleName: string, spanCategory?: string) {
-  const queryConditions: string[] = [`has:${SPAN_DOMAIN} has:${SPAN_DESCRIPTION}`];
-  if (moduleName) {
-    queryConditions.push(`${SPAN_MODULE}:${moduleName}`);
-  }
-
-  if (moduleName === ModuleName.DB) {
-    queryConditions.push(`!${SPAN_OP}:db.redis`);
-  }
-
-  if (spanCategory) {
-    if (spanCategory === NULL_SPAN_CATEGORY) {
-      queryConditions.push(`!has:span.category`);
-    } else if (spanCategory !== 'Other') {
-      queryConditions.push(`span.category:${spanCategory}`);
-    }
-  }
+function getEventView(location: Location, moduleName: ModuleName, spanCategory?: string) {
+  const query = buildEventViewQuery({
+    moduleName,
+    location: {...location, query: omit(location.query, SPAN_DOMAIN)},
+    spanCategory,
+  }).join(' ');
   return EventView.fromNewQueryWithLocation(
     {
       name: '',
       fields: ['span.domain', 'count()'],
       orderby: '-count',
-      query: queryConditions.join(' '),
+      query,
       dataset: DiscoverDatasets.SPANS_METRICS,
       version: 2,
     },
